refactor(core): group graphics extraction helpers
Move image and vector operator-list helpers into a graphics domain folder so the core root keeps fewer low-level extraction modules.
