+++ src/core/graphics/vectorOps.ts
