+++ src/core/processor.ts
@@ -35,7 +35,9 @@ import { buildOutline } from './document/outline.js';
 import { buildPageStructure, countStructureNodes } from './document/structure.js';
 import { buildViewerState, normalizeJavaScriptActions } from './document/viewer.js';
 import { buildFormFields } from './formFields.js';
-import { buildImageBoxes, type ImageOps } from './imageBoxes.js';
+import { buildImageBoxes, type ImageOps } from './graphics/imageBoxes.js';
+import { buildVectorBoxes } from './graphics/vectorBoxes.js';
+import { countVectorPaintOps } from './graphics/vectorOps.js';
 import { buildLayout, markRepeatedBlocks } from './layout.js';
 import { buildLinks } from './links.js';
 import { parsePageRangeWithSkipped } from './pageRange.js';
@@ -52,8 +54,6 @@ import { isRasterBackedTextLayer } from './quality/rasterBackedTextLayer.js';
 import { type CompiledSearch, compileSearch, searchPage, suppressDuplicateOcrMatches } from './search.js';
 import { type JoinItem, joinPageText } from './text/cjkJoin.js';
 import { textMatrixFontSize, textRunGeometryFromTransform } from './textGeometry.js';
-import { buildVectorBoxes } from './vectorBoxes.js';
-import { countVectorPaintOps } from './vectorOps.js';
 import { type BuildVisualRegionsInput, buildVisualRegions } from './visualRegions.js';
 import { detectPageWarnings } from './warnings.js';
 import { extractWidgetAppearanceCaptions } from './widgetAppearance.js';
